backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.3
tensorflow==2.16.1
tensorflow-serving-api==2.16.1
matplotlib==3.8.3
python-jose==3.3.0